+++ docs/components/content/examples/vue/table/ExampleVueTableColumnPinning.vue
@@ -1,0 +1,45 @@
+<script setup lang="ts">
+import type { ColumnDef } from '@tanstack/vue-table'
+import makeData from './makeData'
+import type { Person } from './makeData'
+
+const data = ref(makeData(5))
+
+const columns: ColumnDef<Person>[] = [
+  {
+    header: 'First Name',
+    accessorKey: 'firstName',
+  },
+  {
+    header: 'Last Name',
+    accessorKey: 'lastName',
+  },
+  {
+    header: () => 'Age',
+    accessorKey: 'age',
+  },
+  {
+    header: 'Visits',
+    accessorKey: 'visits',
+  },
+  {
+    header: 'Status',
+    accessorKey: 'status',
+  },
+  {
+    header: 'Profile Progress',
+    accessorKey: 'progress',
+  },
+]
+</script>
+
+<template>
+  <NTable
+    :columns
+    :data
+    :column-pinning="{
+      left: ['status'],
+      right: ['priority'],
+    }"
+  />
+</template>