+++ packages/nuxt/src/module.ts
@@ -122,6 +122,13 @@ export default defineNuxtModule<ModuleOptions>({
     })
 
     addComponentsDir({
+      path: resolve(runtimeDir, 'components/data', 'table'),
+      prefix: options.prefix,
+      global: options.global,
+      watch: nuxt.options.dev,
+    })
+
+    addComponentsDir({
       path: resolve(runtimeDir, 'components/navigation', 'breadcrumb'),
       prefix: options.prefix,
       global: options.global,