+++ packages/nuxt/src/runtime/types/index.ts
@@ -23,6 +23,7 @@ export * from './select'
 export * from './separator'
 export * from './card'
 export * from './breadcrumb'
+export * from './table'
 
 export interface Colors {
   [key: string]: string